requirements.txt
streamlit
pandas
joblib
scikit-learn==1.6.1
huggingface_hub
python-dotenv
numpy
